Biggest Find Yet
So, the reason  why bootstrap was not working that well, was because of MDBootstrap, a very shitty but well designed library. Hence why certain features didn't work. The Front-End Team will have to work tirelessly to fix these issues, but we will pull through and redesign everything from the ground up.