dsfsd
Breeding season - needs calendar with the ability to choose multiple dates so for example 15/3/21 - 28/8/21 on the form, date of birth should be a calendar as well and age should be calculated. Descriptions should have a larger form for input. for est. value, and avg. feed/ day should have blocks to the side and be number only, so the user doesn't automatically put the symbol or put letters in where they shouldn't